Completed work on messaging.js
This has most tests passing.  I can't think of a way to get the auth
tests passing yet, at least in a way that allows for either auth
or no auth db.  Additionally, there are a couple of strange
corner cases I'm still working on solving in two of the tests.