Improve test coverage plus task limit bug
While improving the test coverage and
getting the tests to pass on Travis-CI, I found
a bug with the task-limit code that could cause issue if
there is a password issue.  task-limit is now an
event emitter so it can indicate errors.

background-task has been updated to forward these
on to the clients.