fixes #14 TaskLimit errors are thrown instead of emitted
A precondition check in the TaskLimit.stopTask() method was throwing an exception rather than returning the error in a callback or emitting the error, resulting in unhandled exceptions.  This fix returns it in the callback so it can be handled up the chain.

Fix also addresses tests for this issue, and fixes some other asynchronous handling in the unit tests.