Aborting commit due to empty commit message. Removes a race condition and handles errors returned by TaskLimit.stopTask()
Removes a race condition where stopTask was potentially being called twice, once for the end of the task and once for a task timeout.  This would cause a precondition check to fail in stopTask and throw a
n unhandled exception.

The fix also now passes a callback to stopTask, and if an error is received, emits it.